:fire: :boom: Move some bots stuff to bots.core
