+++ docs/bots/usage.md
@@ -13,14 +13,14 @@ import asyncio
 from vkwave.client.default import AIOHTTPClient
 from vkwave.api.token.token import BotSyncSingleToken, Token
 from vkwave.bots.tokens.storage import TokenStorage
-from vkwave.bots.dispatching.dp.dp import Dispatcher
-from vkwave.bots.dispatching.extensions.longpoll_bot import BotLongpollExtension
-from vkwave.bots.dispatching.router.router import DefaultRouter
+from vkwave.bots.core.dispatching.dp.dp import Dispatcher
+from vkwave.bots.core.dispatching.extensions.longpoll_bot import BotLongpollExtension
+from vkwave.bots.core.dispatching.router.router import DefaultRouter
 from vkwave.bots.tokens.types import GroupId
 from vkwave.api.methods._abc import API
-from vkwave.bots.dispatching.filters.builtin import EventTypeFilter
+from vkwave.bots.core.dispatching.filters.builtin import EventTypeFilter
 from vkwave.longpoll.bot import BotLongpoll, BotLongpollData
-from vkwave.bots.dispatching.events.base import BaseEvent, BotEvent
+from vkwave.bots.core.dispatching.events.base import BaseEvent, BotEvent
 from vkwave.types.bot_events import BotEventType
