+++ vkwave/bots/core/__init__.py
@@ -1,0 +1,4 @@
+from .dispatching.dp.dp import Dispatcher
+from .dispatching.extensions.base import BaseExtension
+from .dispatching.filters.base import BaseFilter
+from .dispatching.filters.builtin import EventTypeFilter, TextFilter
