+++ vkwave/bots/core/dispatching/extensions/base.py
@@ -1,7 +1,7 @@
 import typing
 from abc import ABC, abstractmethod
 
-from vkwave.bots.types.bot_type import BotType
+from vkwave.bots.core.types.bot_type import BotType
 
 if typing.TYPE_CHECKING:
     from ..dp.dp import Dispatcher
