+++ vkwave/bots/core/dispatching/filters/cast.py
@@ -2,7 +2,7 @@ from asyncio import iscoroutinefunction
 from inspect import isawaitable, isfunction
 from typing import Any
 
-from vkwave.bots.dispatching.cast.default import DefaultCaster
+from vkwave.bots.core.dispatching.cast.default import DefaultCaster
 
 from .base import AsyncFuncFilter, BaseFilter, SyncFuncFilter
