+++ vkwave/bots/core/dispatching/handler/cast.py
@@ -1,8 +1,8 @@
 from asyncio import iscoroutinefunction
 from inspect import isawaitable, isfunction
-from typing import Any, cast
+from typing import Any
 
-from vkwave.bots.dispatching.cast.default import DefaultCaster
+from vkwave.bots.core.dispatching.cast.default import DefaultCaster
 
 from .callback import AsyncFuncCallback, BaseCallback, SyncFuncCallback
