+++ vkwave/bots/core/dispatching/handler/record.py
@@ -1,9 +1,9 @@
 from typing import Any, List, Optional, Type, Union, cast
 
-from vkwave.bots.dispatching.filters import EventTypeFilter, filter_caster
-from vkwave.bots.dispatching.filters.base import BaseFilter
-from vkwave.bots.dispatching.handler import BaseHandler, DefaultHandler
-from vkwave.bots.dispatching.handler.callback import BaseCallback
+from vkwave.bots.core.dispatching.filters import EventTypeFilter, filter_caster
+from vkwave.bots.core.dispatching.filters.base import BaseFilter
+from vkwave.bots.core.dispatching.handler import BaseHandler, DefaultHandler
+from vkwave.bots.core.dispatching.handler.callback import BaseCallback
 
 from .cast import caster as callback_caster
