+++ vkwave/bots/core/tokens/strategy.py
