+++ vkwave/bots/core/types/bot_type.py
