+++ vkwave/bots/fsm/filters.py
@@ -1,7 +1,7 @@
-from vkwave.bots.dispatching.events.base import BaseEvent
+from vkwave.bots.core.dispatching.events.base import BaseEvent
 from vkwave.bots.fsm.fsm import State, ForWhat, FiniteStateMachine
 
-from vkwave.bots.dispatching.filters.builtin import BaseFilter, FilterResult
+from vkwave.bots.core.dispatching.filters.builtin import BaseFilter, FilterResult
 from vkwave.bots.storage.types import Key
 
 ANY_STATE = "__any_state__"
