+++ vkwave/bots/utils/uploaders/audio_uploader.py
@@ -1,6 +1,6 @@
 import json
 
-from vkwave.bots.types.json_types import JSONDecoder
+from vkwave.bots.core.types.json_types import JSONDecoder
 from vkwave.http import AbstractHTTPClient
 from vkwave.api.methods import APIOptionsRequestContext
 from vkwave.types.responses import DocsSaveResponseModel
